pyproject.toml
[project]
name = "captcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr==1.7.2",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "streamlit==1.54.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
